+++ src/Repository/PokemonTypeRepository.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace App\Repository;
+
+use App\Entity\PokemonType;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @extends ServiceEntityRepository<PokemonType>
+ *
+ * @method PokemonType|null find($id, $lockMode = null, $lockVersion = null)
+ * @method PokemonType|null findOneBy(array $criteria, array $orderBy = null)
+ * @method PokemonType[]    findAll()
+ * @method PokemonType[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class PokemonTypeRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, PokemonType::class);
+    }
+
+    //    /**
+    //     * @return PokemonType[] Returns an array of PokemonType objects
+    //     */
+    //    public function findByExampleField($value): array
+    //    {
+    //        return $this->createQueryBuilder('p')
+    //            ->andWhere('p.exampleField = :val')
+    //            ->setParameter('val', $value)
+    //            ->orderBy('p.id', 'ASC')
+    //            ->setMaxResults(10)
+    //            ->getQuery()
+    //            ->getResult()
+    //        ;
+    //    }
+
+    //    public function findOneBySomeField($value): ?PokemonType
+    //    {
+    //        return $this->createQueryBuilder('p')
+    //            ->andWhere('p.exampleField = :val')
+    //            ->setParameter('val', $value)
+    //            ->getQuery()
+    //            ->getOneOrNullResult()
+    //        ;
+    //    }
+}