+++ src/Repository/StatRepository.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace App\Repository;
+
+use App\Entity\Stat;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @extends ServiceEntityRepository<Stats>
+ *
+ * @method Stats|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Stats|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Stats[]    findAll()
+ * @method Stats[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class StatRepository extends ServiceEntityRepository
+{
+	public function __construct(ManagerRegistry $registry)
+	{
+		parent::__construct($registry, Stat::class);
+	}
+
+	//    /**
+	//     * @return Stats[] Returns an array of Stats objects
+	//     */
+	//    public function findByExampleField($value): array
+	//    {
+	//        return $this->createQueryBuilder('s')
+	//            ->andWhere('s.exampleField = :val')
+	//            ->setParameter('val', $value)
+	//            ->orderBy('s.id', 'ASC')
+	//            ->setMaxResults(10)
+	//            ->getQuery()
+	//            ->getResult()
+	//        ;
+	//    }
+
+	//    public function findOneBySomeField($value): ?Stats
+	//    {
+	//        return $this->createQueryBuilder('s')
+	//            ->andWhere('s.exampleField = :val')
+	//            ->setParameter('val', $value)
+	//            ->getQuery()
+	//            ->getOneOrNullResult()
+	//        ;
+	//    }
+}