+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.6.0 (2024-03-11)
+
+
+### Features
+
+* add Pokemon stats support ([76376ab](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/76376ab888ea17e1f032a29e38ed40a3d21ffb4e))
+* add tests for Stats ([a373a4c](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a373a4c6d74ceda3b6a3e3a3b724dde6027cebb1))
+* better Pokemon Type structure ([ca3a4de](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/ca3a4de5041a0355f5e2c1639dd945a5f6a49ef5))
+
+
+
 # 0.5.0 (2024-03-10)
 
 ### Features